Redesign score functiong + movement fix

+++ Assets/Scripts/ScoreManager.cs
@@ -7,14 +7,14 @@ public class ScoreManager : MonoBehaviour
 
     [SerializeField] private Text uitext;
 
-    public float score = 0;
-    public void AddScore(float scoreAmount)
+    public int score = 0;
+    public void AddScore(int scoreAmount)
     {
         score += scoreAmount;
         uitext.text = $"Score: {score.ToString()} ";
     }
 
-    public void RemoveScore(float scoreAmount)
+    public void RemoveScore(int scoreAmount)
     {
             score -= scoreAmount;
             uitext.text = $"Score: {score.ToString()} ";
